detect read errors in sniff correctly
we has a signed/unsigned mixup :(

+++ libvips/iofuncs/source.c
@@ -1153,7 +1153,7 @@ vips_source_length( VipsSource *source )
 }
 
 /**
- * vips_source_peek: 
+ * vips_source_sniff_at_most: 
  * @source: peek this source
  * @data: return a pointer to the bytes read here
  * @length: max number of bytes to read
@@ -1219,13 +1219,13 @@ unsigned char *
 vips_source_sniff( VipsSource *source, size_t length )
 {
 	unsigned char *data;
-	size_t bytes_read;
+	gint64 bytes_read;
 
 	if( vips_source_test_features( source ) )
 		return( NULL );
 
 	bytes_read = vips_source_sniff_at_most( source, &data, length );
-	if( bytes_read < length )
+	if( bytes_read < (gint64) length )
 		return( NULL );
 
 	return( data );
